optimize for shotgun WGS

+++ scripts/subgenome_ref_files_1.sh
@@ -5,6 +5,9 @@ if [ -z "$threads" ]; then
 		threads=$((threads-2))
 	fi
 fi
+if [ -z "$lib_type" ]; then
+ lib_type=RRS
+fi
 if [ -z "$nodes" ]; then
  nodes=1
 fi
@@ -35,6 +38,7 @@ fi
 if [ -z "$keep_gVCF" ]; then
 	keep_gVCF=false
 fi
+
 
 
 echo -e "${blue}\n############################################################################## ${yellow}\n- Index Reference Genome \n${blue}##############################################################################${white}\n"
@@ -378,9 +382,8 @@ main () {
 	cd samples
 
 	for i in $( cat ${projdir}/${samples_list} ); do
-		if test ! -f ${projdir}/compress_done.txt && test ! -f ${projdir}/organize_files_done.txt && test ! -f ${projdir}/preprocess/${i%.f*}_redun.sam && test ! -f ${projdir}/preprocess/${i%.f*}_${ref1%.f*}_precall.bam.bai; then
+		if [[ "$lib_type" == "RRS" ]] && test ! -f ${projdir}/compress_done.txt && test ! -f ${projdir}/organize_files_done.txt && test ! -f ${projdir}/preprocess/${i%.f*}_redun.sam && test ! -f ${projdir}/preprocess/${i%.f*}_${ref1%.f*}_precall.bam.bai; then
 			while test ! -f ${i%.f*}_uniq_R1.fasta; do
-				sleep $[ ( $RANDOM % 30 )  + 10 ]s
 				if [[ $(file $i | awk -F' ' '{print $2}') == gzip ]]; then
 					$gunzip -c $i | awk 'NR%2==0' | awk 'NR%2' > ${i%.f*}_uniq.txt 2> /dev/null &&
 					wait
@@ -427,7 +430,7 @@ main () {
 	done
 	wait && touch ${projdir}/organize_files_done.txt
 
-	if [[ "$(wc -l ${projdir}/alignment_summaries/total_read_count.txt | awk '{print $1}')" -le 1 ]]; then
+	if [[ "$lib_type" == "RRS" ]] && [[ "$(wc -l ${projdir}/alignment_summaries/total_read_count.txt | awk '{print $1}')" -le 1 ]]; then
 		cd ${projdir}/alignment_summaries/
 		find -type f -name "*_total_read_count.txt" | xargs cat > total_read_count.hold.txt &&
 		cat total_read_count.txt total_read_count.hold.txt > total_read_count.hold2.txt &&
@@ -438,7 +441,7 @@ main () {
 	wait
 
 
-	if [[ ! -f "${projdir}/align1_${samples_list}" ]] && test ! -f ${projdir}/compress_done.txt; then touch "${projdir}/align1_${samples_list}"; fi
+	if [[ "$lib_type" == "RRS" ]] && [[ ! -f "${projdir}/align1_${samples_list}" ]] && test ! -f ${projdir}/compress_done.txt; then touch "${projdir}/align1_${samples_list}"; fi
 		if [[ "$samples_list" == "samples_list_node_1.txt" ]] && [[ "$mhap_freq" -gt 0 ]] && [[ ! -f "${projdir}/alignment_summaries/background_mutation_test/pop_haps_freqFail.txt" ]]; then
 			align=$(ls ${projdir}/align1_samples_list_node_* | wc -l)
 			while [[ "$align" -lt $nodes ]]; do
@@ -480,7 +483,7 @@ main () {
 	cd ${projdir}/samples
 
 	for i in $(cat ${projdir}/${samples_list} ); do (
-		if test ! -f ${projdir}/hapfilter_done.txt && test ! -f ${projdir}/compress_done.txt && test ! -f "${projdir}/preprocess/${i%.f*}_redun.sam" && test ! -f "${projdir}/preprocess/${i%.f*}_${ref1%.f*}_precall.bam.bai"; then
+		if [[ "$lib_type" == "RRS" ]] && test ! -f ${projdir}/hapfilter_done.txt && test ! -f ${projdir}/compress_done.txt && test ! -f "${projdir}/preprocess/${i%.f*}_redun.sam" && test ! -f "${projdir}/preprocess/${i%.f*}_${ref1%.f*}_precall.bam.bai"; then
 				sleep $[ ( $RANDOM % 30 )  + 10 ]s
 				export nempty=$( ls ${i%.f*}_uniq_R2.fasta 2> /dev/null | wc -l | awk '{print $1}' )
 				 if [[ "$mhap_freq" -gt 0 ]]; then
@@ -523,7 +526,7 @@ main () {
 
 	for i in $(cat ${projdir}/${samples_list} ); do
 		cd ${projdir}/refgenomes
-		if test ! -f ${projdir}/precall_done.txt && test ! -f ${projdir}/preprocess/${i%.f*}_${ref1%.f*}_precall.bam.bai; then
+		if [[ "$lib_type" == "RRS" ]] && test ! -f ${projdir}/precall_done.txt && test ! -f ${projdir}/preprocess/${i%.f*}_${ref1%.f*}_precall.bam.bai; then
 			export nempty=$( ls ${i%.f*}_uniq_R2.fasta 2> /dev/null | wc -l | awk '{print $1}' )
 			while test ! -f ${projdir}/preprocess/${i%.f*}_redun.sam; do
 				if [[ "$nempty" -gt 0 ]]; then
@@ -545,6 +548,26 @@ main () {
 					rm ${projdir}/samples/${i%.f*}_uniq_R1.fasta 2> /dev/null &&
 					wait
 				fi
+			done
+			wait
+		fi
+		cd ${projdir}/samples
+	done
+
+	for i in $(cat ${projdir}/${samples_list} ); do
+		cd ${projdir}/refgenomes
+		if [[ "$lib_type" == "WGS" ]] && test ! -f ${projdir}/precall_done.txt && test ! -f ${projdir}/preprocess/${i%.f*}_${ref1%.f*}_precall.bam.bai; then
+			export nempty=$( ls ${i%.f*}_R2.fastq.gz 2> /dev/null | wc -l | awk '{print $1}' )
+			while test ! -f ${projdir}/preprocess/${i%.f*}_redun.sam; do
+				if [[ "$nempty" -gt 0 ]]; then
+					$java -ea $Xmx2 -cp ${GBSapp_dir}/tools/bbmap/current/ align2.BBMap fast=t qin=33 threads=$threads averagepairdist=$PEdist deterministic=t maxindel=$maxindel local=t keepnames=t maxsites=12 saa=f secondary=t ambiguous=all ref=$ref1 in1=${projdir}/samples/$i in2=${projdir}/samples/${i%.f*}_R2.fastq.gz out=${projdir}/preprocess/${i%.f*}_redun.hold.sam &&
+					wait
+				else
+					$java -ea $Xmx2 -cp ${GBSapp_dir}/tools/bbmap/current/ align2.BBMap fast=t qin=33 threads=$threads maxindel=$maxindel local=t keepnames=t maxsites=12 saa=f secondary=t ambiguous=all ref=$ref1 in1=${projdir}/samples/$i out=${projdir}/preprocess/${i%.f*}_redun.hold.sam &&
+					wait
+				fi
+				mv ${projdir}/preprocess/${i%.f*}_redun.hold.sam ${projdir}/preprocess/${i%.f*}_redun.sam 2> /dev/null &&
+				wait
 			done
 			wait
 		fi
@@ -782,6 +805,11 @@ if [[ "$joint_calling" == false ]]; then
 					Get2_Chromosome=$(awk 'NR>1{print $2,"\t",$3}' ${projdir}/refgenomes/${ref1%.*}.dict | awk '{gsub(/SN:/,"");gsub(/LN:/,""); print $0}' | sort -k2,2 -nr | awk '{print $1}' | awk -v pat=${ref1%.f*} '$0 ~ pat')
 				else
 					Get2_Chromosome=$(echo $Get_Chromosome | tr ',' '\n')
+				fi
+				if [[ ! -z "$Exclude_Chromosome" ]]; then
+					for i in $(echo "$Exclude_Chromosome" | tr ',' '\n'); do
+						Get2_Chromosome=$(echo $Get2_Chromosome | awk -v i=$i '{gsub(i,"");}1')
+					done
 				fi
 				for selchr in $Get2_Chromosome; do
 					$GATK --java-options "$Xmxg -XX:+UseParallelGC -XX:ParallelGCThreads=$gthreads" GenomicsDBImport "${input}" -L "${selchr}" --genomicsdb-workspace-path ${pop}_${ploidy}x_"${selchr}"_raw
